MBL-13582: Streamlined checker, renamed code module.
